requirements.txt
browserhistory == 0.1.2
dotenv == 0.0.5
google-api-core==1.16
google-cloud-language
google-cloud-vision
matplotlib == 3.1.3
opencv_python == 4.2.0.34
pandas == 1.0.1
protobuf == 3.12.2
pymongo == 3.11.2
pysafebrowsing == 0.1.1
